pyproject.toml
[project]
name = "sitegen-agent"
version = "0.1.0"
description = "ADK-based agent that generates a page_plan.json and deploys a Firebase Hosting preview from sitegen-skeleton."
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.34.0",
  "google-auth==2.37.0",
  "requests==2.32.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
